Create features_to_come.md
This file is mainly there for me, its so I can keep track of what features I want to be added to the bot as soon as possible and what features that i want to add to the bot that might take a while. It is also there to show other people what I have planned for the bot.